refactor: :recycle: split files
split two large files (family & member) into several small files for each method
