+++ emr.go
@@ -1,30 +1,47 @@
 package main
 
 import (
-	// "fmt"
 	"log"
+	"time"
 
 	"github.com/renatospaka/emr/domain/entity/family"
-	familyRepository "github.com/renatospaka/emr/infrastructure/repository/family"
+	familyRepository "github.com/renatospaka/emr/infrastructure/repository/family/inMemory"
 )
 
 func main() {
-	famRepo := familyRepository.NewFamilyRepositoryInMemory()
-	fam1 := family.NewFamily("Spakauskas")
-	err := famRepo.Create(fam1)
+	repoFamily := familyRepository.NewFamilyRepositoryInMemory()
+	fam1 := family.NewFamily("Familyname")
+	err := repoFamily.Create(fam1)
 	if err != nil {
-		log.Println("Nova família criada com erro: ", err)
+		log.Println("Nova família criada com erro:", err)
 	} else {
-		log.Println("Nova família criada - ID: ", fam1.ID, ", Surname: ", fam1.Surname)
+		log.Println("Nova família criada -", fam1)
 	}
 
-
-	fam2 := family.NewFamily("")
-	err = famRepo.Create(fam2)
+	repoMember := familyRepository.NewMemberRepositoryInMemory()
+	memb := family.NewMember("Name", "MiddleName", "Lastname", family.Male, time.Date(2006, 8, 5, 0, 0, 0, 0, time.UTC))
+	err = repoMember.Add(memb)
 	if err != nil {
-		log.Println("Nova família criada com erro: ", err)
+		log.Println("Novo membro criado com erro:", err)
 	} else {
-		log.Println("Nova família criada - ID: ", fam2.ID, ", Surname: ", fam2.Surname)
+		log.Println("Novo membro criado! Full name:(", memb.FullName(true), ")")
 	}
 
+	id := memb.ID
+	findMemb, err := repoMember.FindById(id)
+	if err != nil {
+		log.Println(err)
+	} else {
+		log.Println("Achou! Full name:", findMemb.FullName(true))
+	}
+
+	err = repoMember.Remove(id)
+	if err != nil {
+		log.Println(err)
+	}
+
+	err = repoMember.Remove(id)
+	if err != nil {
+		log.Println(err)
+	}
 }
