+++ infrastructure/repository/family/inMemory/memberInMemory.go
@@ -1,0 +1,13 @@
+package familyInMemory
+
+import "github.com/renatospaka/emr/domain/entity/family"
+
+type MemberRepositoryInMemory struct {
+	member []family.Member
+}
+
+func NewMemberRepositoryInMemory() *MemberRepositoryInMemory {
+	return &MemberRepositoryInMemory{
+		member: []family.Member{},
+	}
+}
